requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-cities==0.6.1
django-countries==7.6.1
django-model-utils==4.5.1
djangorestframework==3.15.2
et-xmlfile==1.1.0
idna==3.7
numpy==1.24.4
openpyxl==3.1.4
pandas==2.0.3
psycopg==3.1.19
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
swapper==1.3.0
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
